refactor[next-dace]: move GT4PyWriteBackBufferElimination into its own file
The pass is not part of `gt_simplify()`, so it does not belong into
`simplify.py`. Pure move, no behavioural change.
